+++ pkg/analysis/passes/analysis.go
@@ -22,7 +22,7 @@ import (
 	"github.com/grafana/plugin-validator/pkg/analysis/passes/metadatavalid"
 	"github.com/grafana/plugin-validator/pkg/analysis/passes/modulejs"
 	"github.com/grafana/plugin-validator/pkg/analysis/passes/org"
-	OSVScannerInternal "github.com/grafana/plugin-validator/pkg/analysis/passes/osvscanner-internal"
+	"github.com/grafana/plugin-validator/pkg/analysis/passes/osvscanner"
 	"github.com/grafana/plugin-validator/pkg/analysis/passes/pluginname"
 	"github.com/grafana/plugin-validator/pkg/analysis/passes/published"
 	"github.com/grafana/plugin-validator/pkg/analysis/passes/readme"
@@ -57,8 +57,7 @@ var Analyzers = []*analysis.Analyzer{
 	metadatavalid.Analyzer,
 	modulejs.Analyzer,
 	org.Analyzer,
-	OSVScannerInternal.Analyzer, // new scanner as library
-	// osvscanner.Analyzer,
+	osvscanner.Analyzer,
 	pluginname.Analyzer,
 	published.Analyzer,
 	readme.Analyzer,
