--- pkg/analysis/passes/osvscanner-internal/filter_test.go
@@ -1,18 +1,0 @@
-package OSVScannerInternal
-
-import (
-	"testing"
-)
-
-func TestFilterPackages(t *testing.T) {
-	/*
-		packages := models.VulnerabilityResults{
-			"d3-color",
-			"moment",
-		}
-
-		filteredResults := FilterOSVInternalResults(packages)
-		require.Len(t, filteredResults, 1) // should not have moment
-		require.Equal(t, "d3-color", filteredResults[0])
-	*/
-}
