--- pkg/analysis/passes/osvscanner-internal/testdata/golang/critical/go.mod
@@ -1,43 +1,0 @@
-module github.com/grafana/grafana-kiosk
-
-go 1.19
-
-require (
-	github.com/chromedp/cdproto v0.0.0-20221011223153-490dc4d81f7c
-	github.com/chromedp/chromedp v0.8.6
-	github.com/ilyakaznacheev/cleanenv v1.4.0
-	github.com/smartystreets/goconvey v1.7.2
-)
-
-require (
-	cloud.google.com/go/compute v1.7.0 // indirect
-	github.com/golang/groupcache v0.0.0-20200121045136-8c9f03a8e57e // indirect
-	github.com/golang/protobuf v1.5.2 // indirect
-	github.com/googleapis/enterprise-certificate-proxy v0.1.0 // indirect
-	go.opencensus.io v0.23.0 // indirect
-	golang.org/x/net v0.0.0-20220909164309-bea034e7d591 // indirect
-	golang.org/x/oauth2 v0.0.0-20220822191816-0ebed06d0094 // indirect
-	golang.org/x/text v0.3.7 // indirect
-	google.golang.org/appengine v1.6.7 // indirect
-	google.golang.org/genproto v0.0.0-20220624142145-8cd45d7dbd1f // indirect
-	google.golang.org/grpc v1.47.0 // indirect
-	google.golang.org/protobuf v1.28.0 // indirect
-)
-
-require (
-	github.com/BurntSushi/toml v1.2.0 // indirect
-	github.com/chromedp/sysutil v1.0.0 // indirect
-	github.com/gobwas/httphead v0.1.0 // indirect
-	github.com/gobwas/pool v0.2.1 // indirect
-	github.com/gobwas/ws v1.1.0 // indirect
-	github.com/gopherjs/gopherjs v1.17.2 // indirect
-	github.com/joho/godotenv v1.4.0 // indirect
-	github.com/josharian/intern v1.0.0 // indirect
-	github.com/jtolds/gls v4.20.0+incompatible // indirect
-	github.com/mailru/easyjson v0.7.7 // indirect
-	github.com/smartystreets/assertions v1.13.0 // indirect
-	golang.org/x/sys v0.0.0-20221010170243-090e33056c14 // indirect
-	google.golang.org/api v0.98.0
-	gopkg.in/yaml.v3 v3.0.1 // indirect
-	olympos.io/encoding/edn v0.0.0-20201019073823-d3554ca0b0a3 // indirect
-)
