--- pkg/analysis/passes/osvscanner-internal/testdata/golang/moderate/go.mod
@@ -1,7 +1,0 @@
-module github.com/grafana/grafana
-
-go 1.19
-
-require (
-	github.com/grafana/loki v1.6.2-0.20211015002020-7832783b1caa
-)
