+++ pkg/analysis/passes/osvscanner/lockfile/parse-yarn-lock.go
