+++ pkg/analysis/passes/osvscanner/lockfile/parse-yarn-lock_test.go
