+++ pkg/analysis/passes/osvscanner/lockfile/types.go
