+++ pkg/analysis/passes/osvscanner/testdata/node/critical/package.json
